# contracts/plexor.openapi.yaml
openapi: 3.0.3
info:
  title: Plexor API
  version: 0.1.0-draft
  description: >
    Design-first contract for the Plexor console. Authored on the frontend so
    screens can be built on generated hooks + faker/MSW mocks while the backend
    (Plexor.Host) is not yet available. When the backend lands it implements
    this exact contract; flip VITE_USE_MOCKS=false — screens are untouched.
    Errors follow RFC 7807 Problem Details to match the Platyflow backend convention.
servers:
  - url: /api/v1
    description: Console API (relative — proxied by Vite in dev, real host in prod)
...
paths:
...
  /vsphere/inventory/refresh:
    post:
      operationId: refreshVSphereInventory
      tags: [vsphere]
      summary: Force a fresh vSphere inventory pull
      description: >
        Triggers a synchronous refresh — the call blocks until the
        new snapshot has been written. Returns 202 with the snapshot
        id. 503 when the provider is unconfigured, 502 when vCenter
        is unreachable.
      responses:
        '202':
          description: Refresh succeeded — the new snapshot id is in the body.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VSphereInventoryRefreshResponse' }
        '502':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
...
components:
...
  responses:
    Problem:
      description: RFC 7807 error.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
  schemas:
...
    ProblemDetails:
      type: object
      description: RFC 7807 problem details.
      properties:
        type: { type: string, example: 'https://plexor.dev/problems/not-found' }
        title: { type: string, example: Not Found }
        status: { type: integer, example: 404 }
        detail: { type: string, example: VM not found. }
        instance: { type: string, nullable: true }
...
    VSphereInventoryRefreshResponse:
      type: object
      description: Body for the 202 response on POST /api/v1/vsphere/inventory/refresh.
      required: [snapshotId, status]
      properties:
        snapshotId: { type: string, format: uuid, description: "The newly written snapshot id." }
        status:
          type: string
          enum: [SUCCESS]
          example: SUCCESS
...