# web/apps/console/src/shared/api/src/msw/refreshVSphereInventoryHandler.ts
/* eslint-disable */
// @ts-nocheck
// Generated by kubb — do not edit. Run `bun run generate` from web/tooling/codegen.

/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { RefreshVSphereInventoryMutationResponse, RefreshVSphereInventory502, RefreshVSphereInventory503 } from "../types/RefreshVSphereInventory.ts";
import { http } from "msw";

export function refreshVSphereInventoryHandlerResponse202(data: RefreshVSphereInventoryMutationResponse) {

      return new Response(JSON.stringify(data), {
        status: 202,
          headers: {
          'Content-Type': 'application/json'
        },
      })
}

export function refreshVSphereInventoryHandlerResponse502(data: RefreshVSphereInventory502) {

      return new Response(JSON.stringify(data), {
        status: 502,
          headers: {
          'Content-Type': 'application/problem+json'
        },
      })
}

export function refreshVSphereInventoryHandlerResponse503(data: RefreshVSphereInventory503) {

      return new Response(JSON.stringify(data), {
        status: 503,
          headers: {
          'Content-Type': 'application/problem+json'
        },
      })
}

export function refreshVSphereInventoryHandler(data?: RefreshVSphereInventoryMutationResponse | ((
        info: Parameters<Parameters<typeof http.post>[1]>[0],
      ) => Response | Promise<Response>)) {
  return http.post(`/vsphere/inventory/refresh`, function handler(info) {
      if(typeof data === 'function') return data(info)

      return new Response(JSON.stringify(data), {
        status: 202,
          headers: {
          'Content-Type': 'application/json'
        },
      })
    })
}