# web/apps/console/src/shared/api/src/types/VSphereInventoryRefreshResponse.ts
/* eslint-disable */
// @ts-nocheck
// Generated by kubb — do not edit. Run `bun run generate` from web/tooling/codegen.

/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export const vSphereInventoryRefreshResponseStatusEnum = {
    SUCCESS: "SUCCESS"
} as const;

export type VSphereInventoryRefreshResponseStatusEnumKey = (typeof vSphereInventoryRefreshResponseStatusEnum)[keyof typeof vSphereInventoryRefreshResponseStatusEnum];

/**
 * @description Body for the 202 response on POST /api/v1/vsphere/inventory/refresh.
*/
export type VSphereInventoryRefreshResponse = {
    /**
     * @description The newly written snapshot id.
     * @type string, uuid
    */
    snapshotId: string;
    /**
     * @type string
    */
    status: VSphereInventoryRefreshResponseStatusEnumKey;
};